PXC-2908 : Print proper error message if pxc_encrypt_cluster_traffic is ON and SSL certificates are missing/not matching in the default configuration.
Issue
The error messages, due to a mismatch in SSL configuration, are not very clear or helpful.
This is more important with PXC 8.0 due to the SSL-encryption for all cluster
traffic is now enabled by default:

Solution
Enhance the error text when an SSL is encountered when joining the cluster.
We now display a link to the docs.