PXC-3069: Fixing std::vector assertions
Issue: accessing the address of the first element of an empty vector is
UB, and this is a repeated pattern in galera. This results in crashes
with the centos 8 build.

Fix: changed all uses of [0] to .data() where it makes sense, in the non
test code. Tests and code where we would crash with an empty vector
anyway were left alone.