Improved backup support with the --recv-script option
This changes adds the --recv-script=filename option to garbd.
If this option is specified along with the --sst=... option,
instead of immediately exiting after the SST request is sent,
garbd runs the specified command and waits for the SST request to
complete.

On SST failure, garbd will kill the receiver script and exit with an
exit code.

On successful SST transfer, garbd will wait for the script to finish and
return its return code.