PXC-3031: Handle possible SST hang in garbd
Issue: when multiple (garbd) processes are started around the same time,
there is a good chane that both will request an SST transfer from the
same node at the same time.

The communication protocol doesn't handle this scenario well, and both
processes will think that they correctly selected a donor, and await
the data transfer. However, only one of them will receive it.

Fix: this commit improves the garbd process by sending a SIGINT signal
to the receiver script when the donor signals that the SST process
completed on its side (in practice, data is still in transmit at this
point).
The receiver script should trap this signal, and decide based on this if
it should continue waiting for data, or clean up and exit: if no data
was received until this signal, it should abort.