PXC-3159: Killing the Donor or Connection lost during SST Process Leaves Joiner           Hanging
Problem
-------
When the donor node is killed while SST in progress, the joiner node
hung forever waiting for the SST to complete.

Analysis
--------
In general, if the donor node crashes while SST is in progress and if it
doesn't come back within the timeout decided by the
--sst-initial-timeout option, then SST script aborts with error 32
(Broken pipe) after the timeout is exceeded.

In such cases, Galera was not propagating the error to the above layers,
thereby causing other threads to wait for the state transfer to
complete.

Fix
---
Galera now handles the error by closing the group communication channels
and aborting the server.