PXC-3437: State transfer fails when IST donor select incorrect seqno
Issue: when a joiner requests a state transfer, the donor first checks
if it can serve the transfer using for IST. For this check it prefers the
lowest cluster seqno, which was calculated during the last CC event,
instead of the actual seqno requested by the joiner (if it is smaller).

However it is possible that this seqno was released since then, and it
is no longer present in the galera cache, which causes the state
transfer to fail, with the "Cert index preload first seqno XXX not
found..." error message.

Fix: this commit modifies the min available seqno check so that it also
takes the current local starting GCache entry into calculation.

Since a successful state transfer will result in a configuration change
the actual cc_lowest_seqno value will be corrected after the transfer
completes.