PXC-3774: Relax seqno assertion in replicator_smm
Issue: PXC could crash with this assertion during a 5.7 -> 8.0 upgrade
if no changes were written to the cluster between shutting down the 5.7
node and starting the 8.0 node.

The assertion assumes that the seqno has to increase with the CC event,
but when a node is upgraded equality is also valid in the above
scenario.

Fix: relaxed the assertion to allow for equality, preventing the
5.7->8.0 upgrade from crashing debug buils.