PXC-3907: PXC 8.0.28 refresh - Q1 2022
https://jira.percona.com/browse/PXC-3907

Fixed memory leak detected by AddressSanitizer when running
galera.galera_var_replicate_myisam_on MTR test.

It is possible that progress_ object already exists and is not
deallocated before allocating the new one.
The "normal case" is that it is created when we start moving
DONOR/DESYNCED -> JOINED, and we expect it to be deleted when we move
JOINED -> SYNCED. However we can go back to DONOR/DESYNCED from JOINED
and then when moving again DONOR/DESYNCED -> JOINED we miss one
deallocation.