PXC-3991: Coredumper support for garbd
This commit adds a new "--coredumper=<path>" option to garbd.
When this option is used, and garbd crashes, a coredump will be
written to the given path, with a ".<timestamp" added to it.

The option is only compiled when the WITH_COREDUMPER definition is
set, and its value is 1.

Currently this is only added to the CMake based build, which is
used as a sub-project of the main CMake build, and has the required
libcoredumper dependency set up.