PXC-4298: Make Inconsistency Voting to consider only error codes for voting.
https://jira.percona.com/browse/PXC-4298

Problem
-------

PXC supports Inconsistency Voting Protocol to ensure consistency of
nodes on failure of a DDL. This is done by comparing the error messages
received from other nodes of the cluster and checking if the query
failed with the same error as that of the majority of the cluster to
reach consensus.

This idea, even though works well in general, there are some corner cases
where this logic will fail resulting in cluster breakdown due to
false-positive assumptions.

Some of such examples are listed as below:

Example 1:

Creating indexes in InnoDB is not deterministic in nature. So the same
query can populate indexes in different order due to innodb_ddl_threads
resulting in different errors which scanning the rows.

    node1: Duplicate entry '43-2023-03-18 06:25:17' for key 'iot_devices_values.iot_devices_id_2', Error_code: 1062;
    node2: Duplicate entry '47-2023-05-09 13:21:05' for key 'iot_devices_values.iot_devices_id_2', Error_code: 1062;
    node3: Duplicate entry '43-2023-03-18 06:25:17' for key 'iot_devices_values.iot_devices_id_2', Error_code: 1062;

Example 2:

Different messages are possible when the error message includes the name
of the intermediate temporary table that gets created during the
execution of the ALTER TABLE.

Example 2.1
    node1: Table storage engine for '#sql-f664_33' doesn't have this option, Error_code: 1031;
    node2: Table storage engine for '#sql-f6b4_2' doesn't have this option, Error_code: 1031;
    node3: Table storage engine for '#sql-fbb7_a' doesn't have this option, Error_code: 1031;

Example 2.2

    node1: foreign key constraint fails (`test`.`#sql-25528d_2`, CONSTRAINT `child_ibfk_1`
    node2: foreign key constraint fails (`test`.`#sql-22ed07_9`, CONSTRAINT `child_ibfk_1`

As we can see, in all the above cases even though the query failed with
same error code, the voting protocol considered these messages as
different errors and caused the cluster to go down.

Solution
--------
In order to fix the above problem, this patch modifies the voting
mechanism to consider only the error codes for reaching consensus.

The new logic is:

When a vote message is received from a node, we extract the error codes
from the message using regular expressions, store it in a vector, later
the vector is sorted and stringified as a CSV and the stringified value
of the vector is used for recomputing the vote.

NOTE:
The new logic does not send the recomputed vote to the network. But
instead it sends its vote just like previous versions, but the handling
is done on the receiving end so that backward compatibility is
maintained.