PXC-4343: Corruption of InnoDB file after IST
https://jira.percona.com/browse/PXC-4343

Problem:
After state transfer, in rare cases, when InnoDB starts it fails with
error "Header page contains inconsistent data in datafile", which causes
node to shutdown.
The problem is very hard to reproduce.

Cause:
The problem occurs only if pxc_encrypt_cluster_traffic=ON.
After finishing of IST process, IST receiver closes the socket (which is
in fact the object of class gu::AsioStreamReact) by calling its close()
method. This closes the file descriptor used by the underlying socket.
Then we get to the destructor of AsioStreamReact which shuts down stream
processing engine (and then attempts to close the socket again which is
noop in this case). Shutdown however causes OpenSSL to write to
already closed file descriptor.
If, by chance, another file was opened between 1st call of close() and
the call of shutdown(), and the same file descriptor was assigned to
that new file, shutdown() writes to opened/valid file descriptor, but
not the one we expect. This causes newly opened file corruption.

Solution:
Shutdown stream processing engine before closing underlying socket.
Special credit to Andrew Garner for providing detailed analysis of the
root cause of the problem.